+++ src/Products/MeetingLiege/profiles/liege/import_data.py
@@ -501,9 +501,7 @@ collegeMeeting.powerObservers = (
 
 collegeMeeting.meetingAppDefaultView = 'searchmyitems'
 collegeMeeting.useAdvices = True
-collegeMeeting.usedAdviceTypes = ('positive_finance', 'positive_with_remarks_finance',
-                                  'negative_finance', 'not_required_finance',
-                                  'positive', 'positive_with_remarks', 'negative', 'nil')
+collegeMeeting.usedAdviceTypes = ('positive', 'positive_with_remarks', 'negative', 'nil')
 collegeMeeting.itemAdviceStates = ('itemcreated_waiting_advices',
                                    'proposed_to_internal_reviewer_waiting_advices')
 collegeMeeting.itemAdviceEditStates = ('itemcreated_waiting_advices',
@@ -777,3 +775,14 @@ councilMeeting.category_group_activated_attrs = {
 data = PloneMeetingConfiguration(meetingFolderTitle='Mes séances',
                                  meetingConfigs=(collegeMeeting, councilMeeting),
                                  orgs=orgs)
+data.advisersConfig = (
+    {'advice_types': ['positive_finance',
+                      'positive_with_remarks_finance',
+                      'negative_finance',
+                      'not_required_finance'],
+     'base_wf': 'meetingadviceliege_workflow',
+     'default_advice_type': 'positive_finance',
+     'org_uids': [dfcompta.id, dfcontrol.id],
+     'portal_type': 'meetingadvicefinances',
+     'show_advice_on_final_wf_transition': '1',
+     'wf_adaptations': []},)