+++ src/Products/MeetingLiege/tests/testCustomMeeting.py
@@ -28,8 +28,8 @@ class testCustomMeeting(MeetingLiegeTestCase):
         items = meeting.adapted().getPrintableItemsByCategory()
         itemsWC = meeting.adapted().getPrintableItemsByCategory(groupByCategory=False)
 
-        self.assertEquals(items[0][1], itemsWC[0])
-        self.assertEquals(items[0][2], itemsWC[1])
-        self.assertEquals(items[1][1], itemsWC[2])
-        self.assertEquals(items[1][2], itemsWC[3])
-        self.assertEquals(items[2][1], itemsWC[4])
+        self.assertEqual(items[0][1], itemsWC[0])
+        self.assertEqual(items[0][2], itemsWC[1])
+        self.assertEqual(items[1][1], itemsWC[2])
+        self.assertEqual(items[1][2], itemsWC[3])
+        self.assertEqual(items[2][1], itemsWC[4])